requirements.txt
Django==5.2.4
django-environ==0.12.0
psycopg[binary]==3.2.9
Pillow==11.3.0 